pyproject.toml
[tool.poetry]
name = "cloud-build-badge"
version = "0.1.0"
description = "Badges for Google Cloud Build"
authors = ["Carmine Paolino <carmine@paolino.me>"]
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = "^3.7"
google-cloud-storage = "^2.3.0"
pybadges = "^3.0.0"
PyYAML = "^6.0"
typing-extensions = "^4.2.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.19.0"
pytest = "^7.1.2"
pytest-mock = "^3.7.0"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
